refactor(mapgen): consolidate authoring recipes
