+++ packages/mapgen-core/src/authoring/recipe/index.ts
@@ -1,0 +1,22 @@
+export { deriveRecipeConfigSchema } from "./config-schema.js";
+export { createRecipe } from "./create.js";
+export type {
+  BuildRecipeDagInput,
+  RecipeDag,
+  RecipeDagArtifactRef,
+  RecipeDagDiagnostic,
+  RecipeDagEdge,
+  RecipeDagEndpoint,
+  RecipeDagStage,
+  RecipeDagStageInput,
+  RecipeDagStep,
+  RecipeDagStepContractInput,
+} from "./dag.js";
+export { buildRecipeDag } from "./dag.js";
+export type {
+  CompiledRecipeConfigOf,
+  RecipeConfig,
+  RecipeDefinition,
+  RecipeModule,
+  RecipePublicConfigOf,
+} from "./types.js";
